fix(perfection-p1.a3-fix2): feature-gated clippy-tests cleanup [actual ~1.5h, est 1h, +50%]
Beyond P1 strict scope (default-features clippy --tests). Going broader
to make ALL feature flags clippy-clean per "sempurna" directive.

Default + 13 features fixed:
  default, vulkan, cpp-ffi, smt, gpu, gui, freertos, zephyr, esp32 → 0 errors
  cuda, llvm, native, websocket, mqtt, ble, tls, https, python-ffi → fixed below

Fixes applied:
- src/codegen/llvm/mod.rs: 3 issues (2 PI approx 3.14→1.25 in test
  helpers; 1 always-true assertion removed with rationale)
- src/runtime/gpu/cuda_backend.rs: 1 single-pattern match → if let Ok
- tests/llvm_e2e_tests.rs: missing `no_inline: false` field on FnDef
  literal (latent E0063 since V29.P1 added the field; CI only ran
  feature-less clippy so missed)
- src/codegen/cranelift/runtime_bare.rs: items_after_test_module —
  added #[allow] with rationale (test mod intentionally precedes FFI
  helper extern "C" aliases for locality)
- src/codegen/cranelift/tests.rs: file-level #![allow(approx_constant)]
  (PI/TAU literals are intentional matching .fj source via r#""# blocks)
- src/codegen/cranelift/tests.rs:4721: -1i64 as i64 → -1_i64
- src/interpreter/eval/builtins.rs:10054 + 10141 + 5×websocket/mqtt/ble:
  unneeded `return` statements removed; `match X { (a, b) => ... }`
  destructure-single-pattern → `let (a, b) = X;`
- src/stdlib_v3/net.rs: redundant closure (1 fix via clippy --fix)
- src/ffi_v2/python.rs: PI 3.14 → 1.25 in 2 test sites

Verify gates ALL GREEN:
  cargo clippy --tests --release -- -D warnings → EXIT=0
  cargo clippy --tests --release --features llvm   → EXIT=0
  cargo clippy --tests --release --features cuda   → EXIT=0
  cargo clippy --tests --release --features native → EXIT=0
  cargo clippy --tests --release --features gui    → EXIT=0
  cargo clippy --tests --release --features ble    → EXIT=0
  cargo clippy --tests --release --features mqtt   → EXIT=0
  cargo clippy --tests --release --features websocket → EXIT=0
  cargo clippy --tests --release --features tls    → EXIT=0
  cargo clippy --tests --release --features https  → EXIT=0
  cargo clippy --tests --release --features cpp-ffi → EXIT=0
  cargo clippy --tests --release --features python-ffi → EXIT=0
  cargo clippy --tests --release --features smt    → EXIT=0
  cargo clippy --tests --release --features gpu    → EXIT=0
  cargo clippy --tests --release --features vulkan → EXIT=0
  cargo clippy --tests --release --features freertos → EXIT=0
  cargo clippy --tests --release --features zephyr → EXIT=0
  cargo clippy --tests --release --features esp32  → EXIT=0
  cargo test --lib --release         → 7626 PASS, 0 fail
  cargo test --test '*' --release    → 2501 PASS, 0 fail
  cargo fmt -- --check               → EXIT=0

KNOWN-BROKEN (deferred to P3 B6 feature-matrix audit):
- --features wasm: 5 errors (1 invalid `label: _` syntax + 2 missing
  Default impls + 2 E0063 missing fields on stale Stmt/FnDef literals).
  Latent bugs since fields were added; not in CI matrix.
- --features playground-wasm: 2 errors (E0433 PrettyPrinter doesn't
  exist in formatter mod). Stale API reference.

These features (wasm, playground-wasm) are NOT in CI matrix
(.github/workflows/ci.yml feature-test job runs only llvm/smt/cpp-ffi/
python-ffi). Default-features build is unaffected. P3 (Feature-gate
matrix audit, plan §3 P3) will systematically address full feature-
flag matrix including wasm + playground-wasm fixes.

P1 default-features 100% complete. P1+ "broad scope" 95% complete (2
features in known-deferred).

+++ src/codegen/cranelift/tests.rs
@@ -1,3 +1,10 @@
+// Test-only file: float literals like 3.14 / 6.28 are intentional approximations
+// used as Fajar Lang source code inside r#""# blocks (where Rust clippy lints
+// don't apply to the .fj DSL semantics) AND as matching Rust-side assertions
+// at literal precision. Using std::f64::consts::PI/TAU on the Rust side would
+// silently mismatch the .fj source's 3.14/6.28 literals.
+#![allow(clippy::approx_constant)]
+
 use super::*;
 use crate::codegen::target::TargetConfig;
 use crate::lexer::tokenize;
@@ -4711,7 +4718,7 @@ fn native_string_index_of_not_found() {
             s.index_of("xyz")
         }
     "#;
-    assert_eq!(compile_and_run(src), -1i64 as i64);
+    assert_eq!(compile_and_run(src), -1_i64);
 }
 
 #[test]
